refactor: re-organize code of PrimaryHeader Component
