+++ client/src/layouts/PrimaryHeader/Desktop/SubNavBar.module.scss
@@ -1,19 +1,4 @@
-@use "../../styles/" as s;
-
-// General
-.row-cent {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.col-cent {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
+@use "../../../styles/" as s;
 
 .sub-nav {
   position: relative;
