--- client/src/layouts/PrimaryHeader/PrimaryNav.module.scss
@@ -1,128 +1,0 @@
-@use "../../styles/" as s;
-
-// General
-.row-cent {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.col-cent {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-// Navbar style
-.primary-nav {
-  display: flex;
-  justify-content: center;
-  list-style: none;
-  text-decoration: none;
-  background-color: transparent;
-  margin: 0;
-
-  .nav-item {
-    padding: 1em;
-    font-size: 1.8rem;
-    font-weight: 500;
-    color: var(--dark-black-clr);
-
-    &.active {
-      cursor: default;
-      user-select: none;
-      pointer-events: none;
-      color: var(--primary-clr);
-    }
-
-    &.inactive {
-      &:hover {
-        transition: 0.25s;
-        cursor: pointer;
-        color: var(--primary-clr);
-      }
-    }
-
-    &.nav-item--category:hover .category-section {
-      display: block;
-    }
-  }
-}
-
-// Category style
-.category-section::before {
-  position: absolute;
-  float: left;
-  width: 100%;
-  height: 3em;
-  content: "";
-  display: block;
-  top: -20px;
-}
-
-.category-section {
-  position: absolute;
-  left: 0;
-  z-index: 5;
-  top: 7.2rem;
-  width: 100%;
-  display: none;
-
-  animation: growFromCenter 0.5s;
-
-  .category-wrap {
-    background: var(--dark-black-clr);
-    border-radius: 5px;
-  }
-}
-
-.util-nav {
-  display: flex;
-  flex-direction: column;
-  margin: 1em 0;
-  margin-left: 1em;
-
-  &__header {
-    margin-bottom: 0.5em;
-    padding-bottom: 0.2em;
-
-    font-size: 1.8rem;
-    font-weight: 500;
-    color: var(--dark-white-clr);
-
-    border-bottom: 1px solid var(--dark-white-clr);
-  }
-
-  &__list {
-    color: var(--dark-white-clr);
-    user-select: none;
-    &.upper {
-      margin-top: 0.3em;
-    }
-    &.lower {
-      margin-top: 1em;
-    }
-  }
-
-  &__item {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    width: fit-content;
-    margin-top: 0.3em;
-
-    font-size: 1.6rem;
-    font-weight: 400;
-    color: var(--dark-white-clr);
-    border-bottom: 1px solid var(--dark-black-clr);
-
-    &:hover {
-      width: 100%;
-      transition: 0.5s;
-      color: var(--turquoise-clr);
-      border-bottom: 1px solid var(--turquoise-clr);
-    }
-  }
-}
